+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/DI.kt
@@ -168,6 +168,8 @@ import de.connect2x.messenger.compose.view.room.timeline.element.state.MemberSta
 import de.connect2x.messenger.compose.view.room.timeline.element.state.MemberStateTimelineElementViewImpl
 import de.connect2x.messenger.compose.view.room.timeline.element.state.NameStateTimelineElementView
 import de.connect2x.messenger.compose.view.room.timeline.element.state.NameStateTimelineElementViewImpl
+import de.connect2x.messenger.compose.view.room.timeline.element.state.PowerLevelTimelineElementView
+import de.connect2x.messenger.compose.view.room.timeline.element.state.PowerLevelTimelineElementViewImpl
 import de.connect2x.messenger.compose.view.room.timeline.element.state.TopicStateTimelineElementView
 import de.connect2x.messenger.compose.view.room.timeline.element.state.TopicStateTimelineElementViewImpl
 import de.connect2x.messenger.compose.view.roomlist.RoomListContainerView
@@ -474,6 +476,7 @@ fun timelineViewModule(messengerConfiguration: MatrixMessengerConfiguration?) = 
     timelineElementView<CreateStateTimelineElementView> { CreateStateTimelineElementViewImpl() }
     timelineElementView<EncryptionStateTimelineElementView> { EncryptionStateTimelineElementViewImpl() }
     timelineElementView<HistoryVisibilityStateTimelineElementView> { HistoryVisibilityStateTimelineElementViewImpl() }
+    timelineElementView<PowerLevelTimelineElementView> { PowerLevelTimelineElementViewImpl() }
     timelineElementView<MemberStateTimelineElementView> { MemberStateTimelineElementViewImpl() }
     timelineElementView<NameStateTimelineElementView> { NameStateTimelineElementViewImpl() }
     timelineElementView<TopicStateTimelineElementView> { TopicStateTimelineElementViewImpl() }